feat: add schema for daily-tasks and daily-task-notes
Daily task represents a user's commitments for the day in terms of tasks. These
tasks should be granular enough and achievable in the day. Tasks which aren't
done in a day will be shown in subsequent days as well with a "Delayed by X
days" label

We will represent impediments, challenges, and general comments
on daily-notes as daily-task-notes.